requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
transformers
sentence-transformers
pypdf
python-dotenv
chromadb
faiss-cpu
scikit-learn
torch
safetensors
protobuf
huggingface-hub
accelerate
rank_bm25